# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Car Reservation API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/reservation": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Ambil semua reservasi dengan filter opsional",
                "operationId": "0a552980e5cd60fabc817e71e3124529",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status reservasi",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "payment_status",
                        "in": "query",
                        "description": "Filter berdasarkan status pembayaran",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data reservasi berhasil diambil"
                    },
                    "404": {
                        "description": "Tidak ada data reservasi ditemukan"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Buat reservasi mobil",
                "operationId": "3dab9f4bd788b61ac243ed1e65a3644a",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "Response format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "application/json"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "car_id",
                                    "start_date",
                                    "end_date",
                                    "proof_of_payment",
                                    "payment_status",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "car_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-01"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-05"
                                    },
                                    "proof_of_payment": {
                                        "type": "string",
                                        "example": "bukti-transfer.jpg"
                                    },
                                    "payment_status": {
                                        "type": "string",
                                        "example": "paid"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "pending"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pemesanan Berhasil Dibuat"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reservation/{id}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Ambil detail reservasi berdasarkan ID",
                "operationId": "df313bc0bdfc25bbb75405bb8c148046",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail reservasi ditemukan"
                    },
                    "404": {
                        "description": "Reservasi tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Perbarui status reservasi",
                "operationId": "a9f386ea1bc748af19048475ae1f5748",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_status",
                                    "status"
                                ],
                                "properties": {
                                    "payment_status": {
                                        "type": "string",
                                        "example": "paid"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "confirmed"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status pemesanan berhasil diperbarui"
                    },
                    "404": {
                        "description": "Reservasi tidak ditemukan"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservation"
                ],
                "summary": "Hapus reservasi",
                "operationId": "1f9978e68a9454b50c6727c27b5a84dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservasi berhasil dihapus"
                    },
                    "404": {
                        "description": "Reservasi tidak ditemukan"
                    },
                    "500": {
                        "description": "Kesalahan server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer token authentication using Laravel Sanctum",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "security": [
                {
                    "sanctum": []
                }
            ]
        }
    },
...
}